% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm.tech.rep.R
\name{rm.tech.rep}
\alias{rm.tech.rep}
\title{Removing technical replicates}
\usage{
rm.tech.rep(cntsMerged)
}
\arguments{
\item{cntsMerged}{This is the original expected counts matrix generated from using the rsem.read.iso() function}
}
\description{
This function identifies technical replicates, finds the replicate with the highest reads count, and then remerges these replicates back into the original dataset. The resulting dataset contains expected counts for all strains minus technical replicates, and if a strain had technical replicates the one with the hiest read count is retained.
}
\examples{
rm.tech.rep(cntsMerged)
}
\keyword{replicate}
\keyword{technical}
