% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsem.read.R
\name{rsem.read.iso}
\alias{rsem.read.iso}
\title{Reading in RSEM output - Isoform
This function reads in the RSEM output for each sample and merges the data into one count matrix with rows as transcript IDs and columns as sample names}
\usage{
rsem.read.iso(dir, dirs2, tissue, N)
}
\arguments{
\item{dir}{Global directory. Often points to project start folder. This will be the Parent folder for the folder containing lists of file names for each rsem batch output}

\item{dirs2}{Directory pointing parent folder where RSEM output is}

\item{tissue}{Character value defining name of tissue in file path}

\item{N}{Numerical vector of batch numbers}
}
\description{
Reading in RSEM output - Isoform
This function reads in the RSEM output for each sample and merges the data into one count matrix with rows as transcript IDs and columns as sample names
}
\keyword{rsem}
