% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.content.R
\name{summarize.content}
\alias{summarize.content}
\title{Summarize transcript content into a tabular format}
\usage{
summarize.content(dataset)
}
\arguments{
\item{dataset}{outlist from content.for.plot()}
}
\description{
This takes as input the outlist from content.for.plot()
and creates a table that summarized the number of transcripts
in each category along with totals and percentages.
}
\examples{
summarize.content(content.summary)
}
