% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track.position.table.R
\name{track.position.table}
\alias{track.position.table}
\title{Find Chromosome position for use in UCSC track visualizer}
\usage{
track.position.table(eptOut, ec1 = 25, ec2 = 10, ec3 = 5)
}
\arguments{
\item{eptOut}{output object from exon.tpt.table()}

\item{ec1}{Number of exons associated with transcript 1}

\item{ec2}{Number of exons associated with transcript 2}

\item{ec3}{Number of exons associated with transcript 3}
}
\description{
Brief Description: This function takes as input the output of exon.tpt.table(), and creates a table of chromosome positions to be used
with the UCSC genome browser custome track vizualizer. You can look at 3 different transcripts based on number of exons chosen by the user.
}
\examples{
track.position.table(eptOut, 25, 10, 5)
}
\keyword{Genome}
\keyword{UCSC}
\keyword{browser}
\keyword{custome}
\keyword{track}
